Convert max_scan_size to u64 from usize

+++ src/api.rs
@@ -259,8 +259,9 @@ pub fn fetch_tarball(
     let mut decompressed = GzDecoder::new(response);
     let mut cursor: Cursor<Vec<u8>> = Cursor::new(Vec::new());
     let read = decompressed.read_to_end(cursor.get_mut())?;
-
-    if read > APP_CONFIG.max_scan_size {
+    
+    #[allow(clippy::cast_possible_truncation)]
+    if read > APP_CONFIG.max_scan_size as usize {
         Err(DragonflyError::DownloadTooLarge(download_url.to_string()))
     } else {
         Ok(tar::Archive::new(cursor))
@@ -273,7 +274,8 @@ pub fn fetch_zipfile(http_client: &Client, download_url: &Url) -> Result<ZipType
     let mut cursor = Cursor::new(Vec::new());
     let read = response.read_to_end(cursor.get_mut())?;
 
-    if read > APP_CONFIG.max_scan_size {
+    #[allow(clippy::cast_possible_truncation)]
+    if read > APP_CONFIG.max_scan_size as usize {
         Err(DragonflyError::DownloadTooLarge(download_url.to_string()))
     } else {
         let zip = ZipArchive::new(cursor)?;
